Change the default PATH
The configured default PATH is used in the parts that run as root, the
greeter and for the user  sessions, so its better to have an spartan default.

At the same time, for the zsh users this spartan version gets expanded thanks
to the default global zshenv, while exporting a different value in the
Xsession script would be overriding the users configurations in their local
zshenv.
Forwarded: not-needed
Last-Update: 2015-09-14